Fixes #2450.
Remove xgrid support.  It's broken, we known it's broken, there's
currently no one available to fix it (and no one has been available
for at least a year).  So let's not pretend we have xgrid support; we
don't.  :-(

If anyone would like to fix this situation, please contact us ASAP -
we would love to re-enable xgrid support (all the code is available in
SVN history; it's easy to resurrect).