Fixes #2518: Fix make_tarball -no-ompi in 1.5 branch
r23520
Only build this test if we've enabled the OMPI layer.

r23521
 * Only link in the top-most library that is necessary; it is no
   longer necessary to link in libopen-rte if you link in libmpi (for
   example) because of the fact that libmpi now completely slurps in
   libopen-rte (ditto with libopen-rte and libopen-pal).
 * Only build ompi_rb_tree if we have the OMPI layer.

r23523
Update svn:ignore.

r23524
s/ompi/opal/gi in all of these files because they're really OPAL
tests, not OMPI tests.

This allows us to "make distcheck" with "./autogen.sh -no-ompi"
trees (i.e., these tests will now still work even if the OMPI layer is
not present -- because they're OPAL tests and we should treat them
that way).

r23525
This file should really be distclean, not maintainer clean (it's not
shipped in the tarball).

r23526
Remove the peruse test from any possibility of being run during "make
check" (it's been deactivated for 2+ years now, anyway).  It needs to
be launched via "mpirun" and needs >= 2 processes, so it wasn't a good
candidate for "make check", anyway.

The test itself has moved to OMPI's internal testing suites.

r23527
Move the predefined gap test to ompi/debuggers (we already have the
dlopen_test there, so why not put the other debugger test there with
it?).

r23528
Make it safe to call this script inside of an Open MPI tarball (not
just an SVN or HG checkout).