Fixes #2672: Move MPI_Init_thread C/R fix to v1.4 branch
r24194
MPI_Init_thread is used in place of MPI_Init, so for the checkpoint/restart functionality it must correctly init the C/R functionality instead of simply making a critical section. This allows the C/R thread to be started properly.

Thanks to Takayuki Seki for finding this bug.


r24200
move the INIT to after the error handler, so it matches MPI_INIT. Thanks to Jeff for catching this