Fixes #2641: Terminate after receiving number of sigpipes
We currently just printout a message every time we get a SIGPIPE at mpirun. A user has pointed out that this generates an infinite loop of messages whenever mpirun feeds a pipe and the downstream program exits. Not good.

The attached patch just counts the number of SIGPIPEs we have received and terminates if that number reaches 10. Helps break the infinite loop, while preserving the ability to absorb one caused by some initial startup race conditions.