Fixes #2759: Fix paffinity base MCA duplicate registrations in v1.4
This commit also includes the removal the original registration of
mpi_paffinity_alone (i.e., the non-synonym version that used to be
there -- looks like that got accidentally left in when we converted to
make mpi_paffinity_alone be a synonym of opal_paffinity_alone).

r24569
Fix a bug noted by Gus Correa on the user's list: mpi_paffinity_alone
appeared multiple times in ompi_info output (so did others, but this
is the one that was noticed).  Ensure that we don't repeat
opal_paffinity_base_register_params() multiple times.