Fixes #2785: Fix orterun handling of prefix conflcts
Based on the patch attached to the ticket.

r24657
Check for both absolute-path-to-mpirun and -prefix being specified. If the two differ, print out a warning and ignore -prefix. If they are the same, or only one was given, then proceed as directed.

r24658
Grr...get the prefix right - need to strip the bin out of absolute path to mpirun.