Closes #2830
In 1.4, datatype mismatches in MPI_Accumulate return MPI_ERR_TYPE. But, this was changed in 1.5 and the trunk to return MPI_ERR_ARG. Correspondingly, the onesided test, t_winerr.c, was modified to expect MPI_ERR_ARG when testing for mismatched datatypes, rather than MPI_ERR_TYPE. This, however, resulted in t_winerror test failures with the 1.4 branch. This CMR ensures that MPI_Accumulate in the 1.4 branch returns the same error as does the 1.5 branch and the trunk.