Closes #2831
Oracle would like r24839 brought over to the 1.4 branch. Actually the code that was modified for #2816 is really what we want. I will attach a patch with that code base for 1.4.

We would like to bring this over as soon as reasonable since it stops Oracle from running MTT tests on S11 machines.