Refs #3001.  Still needs a review, but I need to commit this to the v1.5 branch so that George can work on fixing pml v's configure.m4 hackery.
Even in the case where we don't want components to be built, we still
need to run their m4 configry in case they have AM_CONDITIONALs.  This
particular fix is back-ported from the trun.  This allows us to
disable the memory manager, but still run
opal/mca/memory/linux/configure.m4, which contains an AM_CONDITIONAL.
The m4 will still decide ''not'' to build this component, but the
must-always-run-AM_CONDITIONALs criteria will be met.