Fixes #3023: Move attribute fix to v1.6
---svn-pre-commit-ignore-below---
r25965
Wow.  This bug has existed for '''YEARS''' (probably ever since Open
MPI started!).

The FLAG argument to fortran attribute copy functions is a LOGICAL,
meaning that it can only return .TRUE. or .FALSE.  The corresponding C
argument is an int, and the MPI spec says that it must return 1 or 0.
However, in Fortran, .TRUE. is not always necessarily == 1.  So we
need to expand the test to see if it's a Fortran callback.  If so,
check for the Fortran .TRUE. value (not 1).  If it's a C callback,
then check for 1.

r26000
Some updates based on Craig's review of #3023.  Be a little more
defensive about the check of the flag value for the C-based keyvals.
We would never have had a problem because of the specific input data,
but being defensive is good (and it makes the code a little less
subtle / easier to read).

Also add in more comments about exactly what is going on, since this
is complicated stuff.  :-)

r26053
''Really'' fix Fortran attribute copy callbacks.  Put the original
logic back (that was replaced by r25965 and r26000) and fix the one
place that missed OMPI_LOGICAL_2_INT.  This missing OMPI_LOGICAL_2_INT
was the real problem.