Fixes #3068: Move fakeroot ptmalloc pre-main hook fix to v1.6
---svn-pre-commit-ignore-below---
r26203
Fixes #2812: check for env. markers indicating that we're in a
fakeroot.  If so, exit out of the pre-main hook immediately (without
calling functions such as stat, which will be replaced by fakeroot to
things that are not safe to call in a pre-main environment).

r26204
Using the right env variable name helps!