Fixes #3293: Cleanup hostfile behavior in managed environments
Custom patch from the ticket:

When working in a managed environment, we want any hostfile
specifications to be met, but error out if any hosts in the hostfile
are not part of the allocation. We also want to ensure that we
properly assemble the hosts in an unmanaged allocation.

This patch brings over changes from the trunk that update the
allocation logic, removes the default hostfile from filtering of
managed allocations, and generates an error message + exit if a
hostfile contains a host that isn't in a managed allocation.