Fixes #3521: Move r28102 to v1.7 branch
---svn-pre-commit-ignore-below---

r28102 [[BR]]
Thanks to Brian for noticing an incorrect MPI_ERR value in the Fortran
MPI constants, which led to also find that MPI_ERR_WIN was mistakenly
not defined in the Fortran MPI constants.

This fix also needs to be applied to v1.7.  Brian -- can you review?

cmr:v1.7