Fixes #3632: Move <framework>_base_verbose fix to v1.6
Custom patch from ticket:

The problem was essentially that we had an ambiguity in
mca_base_components_close().

This function is called to close a list of components. It's usually
called during the close of the entire framework, but sometimes it's
called during framework initialization or selection, when we have
generated a list of components that don't want to be used (e.g., if
the openib BTL decides that it doesn't want to be used, the BTL base
adds it to the to_be_closed list, and then calls
mca_base_components_close() with that list near the end of BTL
selection).

The mca_base_components_close() function was closing the output stream
in most cases. Bad!

I added a explicit bool param to mca_base_componets_close() that
indicates whether the caller wants to close the stream or not. This
change touched just about every framework (e.g., in their close
functions) :-(, but it seemed like the Right Way to fix it.