Fixes #3662: Move r28753 to v1.7 branch (hcoll fix)
---svn-pre-commit-ignore-below---

r28753 [[BR]]
This fixes the nasty configure.m4 hack that was added long ago and not removed. My fault for not catching earlier. I've also removed the '.ompi_ignore' in coll/hcoll. Throwing this to Nathan for review. Upon successful review, this should be added to cmr:v1.7:reviewer=hjelmn