Fixes #3676: Move opal/if and hotel improvements to v1.7
---svn-pre-commit-ignore-below---

r28343 [[BR]]
Defensive programming: add a constructor for opal_if_t that zeros
everything out before using it.  

This is not in response to any known bug, but rather just a
pre-emptive, defensive move to help prevent bugs in code that forgets
to initialize a field.

r28344 [[BR]]
Also support getting the MAC and MTU.

r28345 [[BR]]
Add public functions for retrieving the MAC and MTU (paired with
r28344).

r28395 [[BR]]
Ahem....cleanup a copy/paste error in naming of these functions

r28757 [[BR]]
Add an oh-so-slightly faster variant of the hotel "checkin" action
(since this is used in the fast path) for when you ''know'' that there
will be a room available:

 * Don't do the last_unoccupied_room check
 * Return void