Fixes #3657: Check all addresses in ifislocal
---svn-pre-commit-ignore-below---

Custom patch from ticket based on:

r28716 [[BR]]
If we enable ipv6, we resolve a hosts addresses and check them all against our local interfaces to determine if the given host is us. However, if we don't enable ipv6, we only checked the first address returned. This can cause us to incorrectly identify a hostname as "not us".

Make -disable-ipv6 behave the same as --enable-ipv6 by checking all the returned addresses.