Fixes #3719: Move r29013 to v1.7 branch (Fix debugger support for direct-launched)
---svn-pre-commit-ignore-below---

r29013 [[BR]]
Fix debugger support for direct-launched jobs.

The orte rte component checks the orte_standalone_operation to decide
if it should wait for a message from the hnp or wait on the debugger.
This variable needed to be set to true in ess/pmi to enable the
correct path when direct launching.

cmr=v1.7.3:reviewer=rhc
cmr=v1.6.6:reviewer=rhc