Fixes #3725: Move r29041, 29043, r29068 to v1.7 branch
---svn-pre-commit-ignore-below---

r29041 [[BR]]
Fix --disable-mpi-profile option so it can build

cmr:v1.7.3:reviewer=hjelmn


r29043 [[BR]]
Okay, fix it so it works both --disable-mpi-profile and --enable-mpi-profile. I'm not sure why mpit's library has to be treated differently, but it seems that it needs some special care to work in both scenarios

Refs #3725


r29068 [[BR]]
Fix compilation of the MPI tools information interface when profiling
is enabled and fix a bug in the handling of watermark performance
variables.

cmr=v1.7.3:ticket=#3725:reviewer=jsquyres