Fixes #3734: Move r29070, r29078 to v1.7 branch (ompi_info improvements w.r.t. new MCA/MPI_T system)
---svn-pre-commit-ignore-below---

r29080 [[BR]]
As per the discussion on the devel telecon, do not compute ompi_comm_world_thread_level_mult if thread multiple is disabled. We aren't using the value anyway, but we will leave the current code in-place until we understand if it is needed or not.



r29078 [[BR]]
ompi_info: mark the variables from disabled components as disabled in
the output of ompi_info.

A variable is disabled if its component will never be selected due to
a component selection parameter (eg. -mca btl self). The old behavior
of ompi_info was to not print these parameters at all. Now we print the
parameters. After some discussion with George it was decided that there
needed to be some way to see what parameters will not be used. This was
the comprimise.

This commit also fixes a bug and a typo in the pvar sytem. The enum_count
value in mca_base_pvar_dump was being used without being set. The full_name
in mca_base_pvar_t was not being used.

cmr=v1.7.3:ticket=#3734