Fixes #3737: Move r29086, r29101 to v1.7.3 (Cleanup PMI minimization)
---svn-pre-commit-ignore-below---

r29086 [[BR]]
We don't need to report errors in pmi_setup as it is possible that PMI is available, but that we weren't launched under it (e.g., we launched via mpirun).

cmr:v1.7.3:reviewer=hjelmn:subject="Silence unnecessary PMI error msgs"


r29101 [[BR]]
Fix compile error in the heterogeneous case.

We've been forcing C99 compiler compliance for a while now, so use C99
syntax to keep the #if code tidy.