Fixes #3767: Move r29169 to v1.7 branch (The function orte_iof_base_setup_prefork attempts to)
---svn-pre-commit-ignore-below---

r29169 [[BR]]
The function orte_iof_base_setup_prefork attempts to create a pty for
child stdout and falls back to plain pipe if openpty fails. Child uses
the 'usepty' flag to decide whether to treat this descriptor as a pty
or as a pipe.
Set 'usepty' flag to 0 upon openpty failure to inform the child that
it isn't dealing with a pty even though pty has been requested.


Thanks to Michal Peclo for reporting it and providing a patch.

cmr:v1.7.3:reviewer=jsquyres
cmr:v1.6.6:reviewer=jsquyres