Fixes #3775: Move r29201 to v1.7 branch (Silence a warning about an)
---svn-pre-commit-ignore-below---

r29201 [[BR]]
Silence a warning about an address that can never be NULL - libevent needs to deal with the situation where the user may have compiled the code on a system where this function is present, but executes it on one where it isn't. Thus, a compile-time test isn't adequate.

Pushed upstream.

cmr:v1.7.3:reviewer=jsquyres


r29210 [[BR]]
C standard stipulates that we have to cast the function to another of the same type to avoid unexpected behavior. We aren't using the function in this case, but Nick correctly points out that we should follow the standard regardless.

Refs #3755