Fixes #3672: Move r28682 to v1.7 (topo files for mapping purposes)
---svn-pre-commit-ignore-below---

r28682 [[BR]]
There are cases where we want to use the novm state machine, but the backend node topology differs from that where mpirun is executing. In those cases, we can wind up thinking we are oversubscribed because the head node has fewer cores than the compute nodes.

To resolve this situation, add the ability to specify a backend topology file that mpirun shall use for its mapping operations. Create a new "set_topology" function in opal hwloc to support it.