Fixes #3811: Move r29285 to v1.7 branch (fix convertor set_position)
---svn-pre-commit-ignore-below---

r29285 [[BR]]
Fix a corner case for a non-contiguous send convertor where the
convertor accepted to be set to a position in the middle of a
predefined datatype. Once set there is was unable to provide the
second part of the datatype. This fix force the convertor to be
aligned on predefined datatypes boundaries for any non-contiguous
send convertor.

r29296 [[BR]]
regression test for r29285 (convertor_set_position)