Fixes #3824: Move r29341 to v1.7.3 (Restrict PMI db component)
---svn-pre-commit-ignore-below---

r29341 [[BR]]
Fix an issue regarding use of PMI when running processes and tools that don't need or want to use it. We build PMI support based on configuration settings and library availability.
 However, tools such as mpirun don't need it, and definitely shouldn't be using it. Ditto for procs launched by mpirun.

We used to have a way of dealing with this - we had the PMI component check to see if the process was the HNP or was launched by an HNP. Sadly, moving the OPAL db framework removed
 that ability as OPAL has no notion of HNPs or proc type.

So add a boolean flag to the db_base_select API that allows us to restrict selection to "local" components. This gives the PMI component the ability to reject itself as required. W
e then need to pass that param into the ess_base_std_app call so it can pass it all down.


r29348 [[BR]]
Fix typo in grpcomm_pmi_module.c that was giving the wrong locality for direct launched jobs. Refs #3824