Fixes #2715: MPI-3: Non-blocking collectives
---svn-pre-commit-ignore-below---

r29333 [[BR]]
Add basic algorithms for the remaining non-blocking collectives.

The algorithms are intended for MPI-3.0 compliance and are not
optimized. We should aim to add better algorithms in the future through
cheetah.

MPI_Iallreduce and MPI_Igatherv on intercommunicators are required for
MPI_Comm_idup support.

cmr=v1.7.4:reviewer=brbarret:ticket=#2715