Fixes #3903: Move r29661 to v1.7 branch (Revise the launch timeout detection)
---svn-pre-commit-ignore-below---

r29661 [[BR]]
Revise the launch timeout detection so we don't mistakenly declare "failed to start". Recognize that timeout is at the per-job level, and define the timeout param as a total value instead of seconds/daemon as it otherwise can get to be an enormous (and useless) number.

Resolves problems in loop_spawn where the timer was incorrectly firing and killing the overall job.

cmr=v1.7.4:reviewer=hjelmn