Fixes #3901: Minor usnic MPI_T updates
---svn-pre-commit-ignore-below---

r29656 [[BR]]
Expand the "btl_usnic" MPI_T enumeration to have strings of the form:

   <usnic device name>,<eth device>,<ip address>/<CIDR prefix>

For example:

   usnic_0,eth4,10.1.0.15/16

This is just handy for mapping the usnic_X device back to the IP
network to which it corresponds.

r29657 [[BR]]
Arrgh -- remove debugging printf.

r29664 [[BR]]
Refs #3901 

Feedback from Dave's review.

r29665 [[BR]]
Refs #3091

Gah!  The "device" variable isn't used at all in this loop (my eye
glossed over the next line and thought that "device" was used in the
free() statement, but it's actually "devices" -- not "device").