Fixes #3909: Move r29681 to v1.7 branch (update -wno-long-double test to support clang)
---svn-pre-commit-ignore-below---

r29681 [[BR]]
Update -Wno-long-double test to support clang.

This prevents bazillions of warnings from clang tha -Wno-long-double
isn't supported.

The warning that clang issues when it see -Wno-long-double is does not
use any of the words we were looking for, so I added "unknown" to the
list of words to look for.  I also re-indented the two m4 tests so
that they're a bit more readable.

cmr=v1.7.4:reviewer=brbarret:subject=Update -Wno-long-double test to support clang