Fixes #3966: Move PORT_ACTIVE usnic fix to v1.7
---svn-pre-commit-ignore-below---

r29852 [[BR]]
Technically, the PORT_ACTIVE is not a "bad" event.

Note that this event should never happen within a single OMPI job,
because OMPI will ignore usnic ports that are down.  The PORT_ACTIVE
event should only occur if a port ''was'' down and is now ''up''.  But
what the heck -- if we ever do get this event, it is harmless -- just
ignore it.