Fixes #4004: Move r29991 to v1.7 branch (fix sparc sigbus in opal convertor code)
---svn-pre-commit-ignore-below---

r29991 [[BR]]
Fixes #3989: opal_pack_debug was instantiated as a bool in one file
and extern'ed s an int in another.  This caused a SIBGUS on
Solaris/SPARC.

This commit properly moves the extern to a .h file so that it's the
same in all files.  It also moves the DO_DEBUG to the header file,
because it was defined to the same thing in multiple .c files.

cmr=v1.7.4:reviewer=bosilca:subject=fix SPARC SIGBUS in opal convertor code

r30018 [[BR]]
Cleanup all the datatype test to avoid any memory leaks or RUI from valgrind.

r30019 [[BR]]
Correctly initialize and finalize all the datatype classes. No memory leaks on the
datatype engine subsists.

r30020 [[BR]]
rearrange the fields to remove a gap in the datatype.

r30021 [[BR]]
Remove unused variable

Refs #4004