Fixes #4019: Move r30042 to v1.7 branch (remove an annoying warning)
---svn-pre-commit-ignore-below---

r30042 [[BR]]
Remove an annoying warning. If the user excludes a non-existent interface, there is no reason to warn - the interface may simply not exist on that node.

cmr=v1.7.4:reviewer=jsquyres:subject=Remove an annoying warning


r30091 [[BR]]
As Jeff pointed out, the reqd flag should only turn off the show_help - still enter the rest of the code block

Refs #4019


r30146 [[BR]]
Per 1/7/2014 telecon: Add an MCA param to turn on all warnings for missing excluded interfaces.

Refs #4019


r30161 [[BR]]
Change the MCA level to 2 (from 5) on the rationale that it may be
needed for correctness.  The if_include/if_exclude are level 1, and
the TCP port range params are level 2; this parameter seems to be on
par with the TCP port range params.

Refs #4019