Fixes #4073: Move r30179 to v1.7 branch (do a better job of reporting bindings)
---svn-pre-commit-ignore-below---

r30179 [[BR]]
Do a better job of reporting bindings - if someone gives a spec that binds us to all processors, then we are effectively unbound and should report it clearly instead of outputting a long line of B's.

cmr=v1.7.4:reviewer=jsquyres:subject=Do a better job of reporting bindings


r30204 [[BR]]
Protect against orte_proc_applied_binding being NULL.

It is now possible for orte_proc_applied_binding to be NULL (e.g., if
you mpirun --bind-to none), so we need to ensure we don't pass it down
to opal_hwloc_base_cset2*str().

Also, take the opprotunity to de-duplicate some strings that are used
in multiple places.

Refs #4073