Fixes #4113: Move r30289 to v1.7 branch (udreg: free huge page allocations)
---svn-pre-commit-ignore-below---

r30289 [[BR]]
udreg: free huge page allocations correctly

This commit fixes an error path that occurs when huge page allocations are
enabled. In this case we allocate a huge page and try to register it but fail.
We then were calling free on the opal object. Fix this by calling the proper free
function.

cmr=v1.7.4:reviewer=rhc