Fixes #4122: Move r30302 to v1.7 branch (fix mpicxx wrapper compiler)
---svn-pre-commit-ignore-below---

r30302 [[BR]]
Remove an optimization to not bother setting up the C++ compiler if we
weren't building the MPI C++ bindings because this causes "CXX=no" to
be passed to the mpicxx wrapper compiler.  

Unfortunately, this optimization was brought to the v1.7 branch, and
therefore if you --disable-mpi-cxx on the v1.7 branch, the same
problem will happen.  So this needs to be CMR'ed to v1.7 as well.

Submitted by Jeff, reviewed by Brian.  RM approved by Brian.

Fixes #4120.

cmr=v1.7.4:reviewer=ompi-gk1.7:subject=Fix mpicxx wrapper compiler