Fixes #4150: Move r30348 to v1.7 branch (fix mips assembly)
---svn-pre-commit-ignore-below---

r30348 [[BR]]
Apply a patch provided by Paul Hargrove back in Jan 2013 that fixes MIPS assembly issues. This patch was originally reviewed and RM-approved to go into the 1.6 branch (which never happened for logistical reasons), and subsequently the trunk patch was provided. Paul has verified the patch and its application to 1.7.4, so we will consider it reviewed for that purpose.

cmr=v1.7.4:reviewer=ompi-rm1.7:subject=Fix MIPS assembly