Fixes #4169: Move r30397 to v1.7 branch (This is an old patch)
---svn-pre-commit-ignore-below---

r30397 [[BR]]
This is an old patch (r26226) from two years ago that somehow went directly into the 1.6 branch without first entering the trunk. Hence, the problem it fixed still remains in the trunk, and in the 1.7 series as a regression :-(

Thanks to Paul Hargrove for tracking it down.

RM-approved

cmr=v1.7.4:reviewer=ompi-gk1.7