Fixes #4198: Move r30475 to v1.7 branch (Fix suicide operation when MPI)
---svn-pre-commit-ignore-below---

r30475 [[BR]]
Fix suicide operation when MPI app loses connection to its local daemon. In that scenario, we correctly callback up to the MPI layer notifying it of the lost connection. However, when the MPI layer calls back down to tell the RTE to abort, it is passing back a flag indicating we should report that error to our local daemon - which is dead. This leads to an infinite loop. Break it by using checking the flag indicating an abnormal term was ordered by the RTE and thus don't attempt to send the message.

cmr=v1.7.4:reviewer=jsquyres