Fixes #4211: Move r30524 to v1.7 branch (Ensure configure test compiler turds)
---svn-pre-commit-ignore-below---

r30524 [[BR]]
Ensure configure test compiler turds are removed.

This particular test can leave various <foo>.mod files in the OMPI top
build dir, which leads developers to say "what are these .mod files in
my 'svn status' output?".

So make a subdir, run the tests in there, and then remove the subdir.

Reviewed by Dave Goodell.

This is not 100% necessary for OMPI 1.7.4, but it would be nice.  The
goal is to have this test removed by OMPI 1.7.5 (i.e., fix ticket
4157).  So I leave it up to the RM to decide whether this goes into
1.7.4 or not.

cmr=v1.7.4:reviewer=ompi-rm1.7