Fixes #4174: Move r30401 to v1.7 branch (Fix ia64 operations by correcting)
---svn-pre-commit-ignore-below---

r30401 [[BR]]
Fix ia64 operations by correcting a couple of bugs in the ia64 atomics. Thanks to Paul Hargrove for the patch!

Since Paul is the only one of the team with the required hardware to test it, and he has done so, consider this RM-approved.

cmr=v1.7.5:reviewer=ompi-gk1.7