Fixes #4241: Move r30644 to v1.7 branch (OPENIB: Fixed error from writing)
---svn-pre-commit-ignore-below---

r30644 [[BR]]
OPENIB: Fixed error from writing to an uninitialized pipe. 
The error was caused by leaving the pipe to the async thread uninitialized, then writing to it regardless of this. 
Fix is to check the existance of the async thread and the pipe to it.

reviewd by miked

cmr=v1.7.5:reviewer=ompi-rm1.7