Fixes #4239: Move r30642 to v1.7 branch (purge stale session dirs at startup)
---svn-pre-commit-ignore-below---

r30642 [[BR]]
In case there are stale session directories around, do a purge of the relevant session directory tree when an orted, HNP, or singleton start. This won't help in the case of direct-launched apps, but it's the best we can do.

cmr=v1.7.5:reviewer=jsquyres:subject=purge stale session dirs at startup