Fixes #4267: Move r30719 to v1.7 branch (handle non-matching addresses)
---svn-pre-commit-ignore-below---

r30719 [[BR]]
If we are given a TCP oob address that doesn't match any active module, it is still possible that we could route to the address if a router is in the system. No harm in trying, so arbitrarily pick the first connection in the active module list and assign the peer to it. If that module can't reach it, we'll follow the usual failover mechanism until finally concluding that nobody can get there.

cmr=v1.7.5:reviewer=jsquyres:subject=handle non-matching addresses


r30733 [[BR]]
Make Jeff smile - pretty-up the indentation

Refs #4267


r30740 [[BR]]
Fix bugs in the oob base - ensure we get the components in high-to-low priority, and that we correctly track reachability via all components. Adjust the priority of the tcp component to leave headroom for others

Refs #267