Fixes #4286: Move r30760 to v1.7 branch (OSHMEM: Fixed race condition problem)
---svn-pre-commit-ignore-below---

r30760 [[BR]]
OSHMEM: Fixed race condition problem in example code. 
The problem was with oshmem ring code, which cycles a constant amout of data through all the processes.
When the program would expect the memory to change it did not explicitly warrent a recv() call, thus
counting on the memory region to hold the new value when in fact it could be set before or after the
check, causing slower BTLs to pass and faster ones to fail. The fix changes the logic to anticipate
the next message rather then the current one, which would be a mistake.
Patch applied to both C and fortran90 version of the example code.

reviewd by miked

cmr=v1.7.5:reviewer=ompi-rm1.7