Fixes #4387: Move r31061 to v1.8 branch (Fix some set-but-not-used compiler warnings.)
---svn-pre-commit-ignore-below---

r31081 [[BR]]
coll/ml: fix a case not correctly handled by r31071

In r31071 I modified the logic to not increment the hierarchy level if
no processes were selected by that sbgp. That fixed a problem seen on
systems where we don't support process binding. The problem is there
is a case where we actually did select processes yet the number of
selected processes is 0. We need to increment the hierarchy in this case
as well.

This should fix the segmentation fault found by recent MTT runs. Once
this is committed to 1.7.5 remove the .ompi_ignore's from coll/ml and
bcol/ptpcoll. Tested with ompi-tests/ibm.

cmr=v1.7.5:reviewer=rhc