Fixes #4393: Move r31068 to v1.7 branch (Deal with the corner case)
---svn-pre-commit-ignore-below---

r31068 [[BR]]
Deal with the corner case where we encounter an error when attempting to launch a daemon. In this case, we will order abnormal termination before daemons callback to us, and thus any attempt to send them a "die" message will fail. Ensure that mpirun at least exits cleanly in this scenario, thereby allowing the remote daemons that did get launched to commit suicide when comm fails.

cmr=v1.7.5:reviewer=jsquyres


r31080 [[BR]]
Remove a check that prevented mpirun from exiting when it should in the single-node case

Refs #4393