Fixes #4441: Move r31197, r31223 to v1.8 branch (osc/rdma: quiet warning in r31197)
---svn-pre-commit-ignore-below---

r31197 [[BR]]
osc/rdma: add support for datatypes with large descriptions

This commit adds large datatype description support to the osc/rdma
component. Support is provided by an additional send/recv of the datatype
description if the description does not fit in an eager buffer. The
code is designed to require minimal new code and not for speed. We
consider this code path to be a slow path.

Refs #1905

cmr=v1.8:reviewer=jsquyres

r31223 [[BR]]
osc/rdma: quiet warning in r31197

cmr=v1.8:ticket=#4441