Fixes #4480: Move r31288 to v1.8 branch (Ensure that orted's are not)
---svn-pre-commit-ignore-below---

r31288 [[BR]]
Ensure that orted's are not bound to single processors if the TaskAffinity option is set by default. Thanks to Artem Polyakov for the patch, and for his patience in explaining the situation.

Reviewed with Moe Jette to ensure this was correct, and confirmed by me.

RM-approved

cmr=v1.8:reviewer=ompi-gk1.8