Fixes #4511: Move darray fix to v1.8
---svn-pre-commit-ignore-below---

r26862 [[BR]]
Fix the darray issue where the UB was not computed correctly. The
old version of the code tried to use the MPI_UB marker, but this
failed if the old marker (the one set in the cyclic function) had
a larger value. Replace the hardcore markers MPI_LB and MPI_UB by
their softer counterparts (using the _resize function).